backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytesseract
pillow
psycopg2-binary
python-dotenv